src/toron/_data_access/property_repository.py
"""PropertyRepository and related objects using SQLite."""

import sqlite3
from json import dumps as json_dumps

from .base_classes import BasePropertyRepository, JsonTypes


class PropertyRepository(BasePropertyRepository):
    def __init__(self, data_reader: sqlite3.Cursor) -> None:
        """Initialize a new PropertyRepository instance."""
        self._cursor = data_reader

    def add(self, key: str, value: JsonTypes) -> None:
        """Add an item to the repository."""
        self._cursor.execute(
            'INSERT INTO main.property (key, value) VALUES (?, ?)',
            (key, json_dumps(value)),
        )

    def get(self, key: str) -> JsonTypes:
        """Retrieve an item from the repository."""
        self._cursor.execute(
            'SELECT value FROM main.property WHERE key=?',
            (key,),
        )
        result = self._cursor.fetchone()
        if result:
            return result[0]
        return None

    def update(self, key: str, value: JsonTypes) -> None:
        """Update an item in the repository."""
        self._cursor.execute(
            'UPDATE main.property SET value=? WHERE key=?',
            (json_dumps(value), key),
        )

    def delete(self, key: str):
        """Remove an item from the repository."""
        self._cursor.execute(
            'DELETE FROM main.property WHERE key=?',
            (key,),
        )
